requirements.txt
cryptography==41.0.4
jwt==1.3.1
pandas==2.1.1
py==1.11.0
PyJWT==2.8.0
requests==2.31.0
toml==0.10.2
urllib3==2.0.5
pip==24.0
wheel==0.41.2
setuptools==68.2.2
google-cloud-resource-manager==1.10.4
ipwhois==1.2.0